move sort functions to sort.js
