+++ js/listpage/addListEventListeners.js
@@ -1,3 +1,5 @@
 import { addEventHandlers } from "./listEventListener";
+import { addSortUIAndListeners } from "./list_sort/sort";
 
 addEventHandlers();
+addSortUIAndListeners()
